fixes and smaller refactorings
