+++ src/video.rs
@@ -2,8 +2,7 @@ use std::borrow::Borrow;
 use std::convert::AsRef;
 use std::string::ToString;
 use std::error::Error;
-use std::fmt;
-use api::{AlbumId, Bool, Collection, Duration, FullId, Id, LikesCount, OwnerId, ReportReason, Sort, Timestamp, SortOrder};
+use api::{AlbumId, Bool, Collection, Duration, FullId, Id, LikesCount, OwnerId, ReportReason, Sort, Timestamp, SortOrder, Comment, Attachment};
 use serde::de::Deserialize;
 
 #[cfg(feature = "unstable")]
@@ -396,26 +395,6 @@ request! {
     #[derive(Eq, Copy)]
     struct HideCatalogSection for ["video.hideCatalogSection"](v => 5.44) -> Bool {
         section_id: Id = () => {},
-    }
-}
-
-enum_str! { AttachmentKind {
-    Photo = "photo",
-    Video = "video",
-    Audio = "audio",
-    Document = "doc",
-}}
-
-#[derive(Eq, Copy, Clone, PartialEq, Debug)]
-pub struct Attachment {
-    pub kind: AttachmentKind,
-    pub owner_id: OwnerId,
-    pub media_id: Id,
-}
-
-impl fmt::Display for Attachment {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}{}_{}", self.kind.as_ref(), self.owner_id, self.media_id)
     }
 }
